Changes to file structure
1. Moved Express.ts to providers
2. Added App.ts to bootstrap providers.
3. Removed js files from build.
4. Fixed tab indentation in tslint
5. Added files in .gitignore.
