+++ pmt-auth/.gitignore
@@ -1,3 +1,5 @@
 node_modules/
 .env
-dist
+build
+npm-debug.log
+.DS_Store
