+++ pmt-auth/tslint.json
@@ -8,7 +8,8 @@
       ],
       "indent": [
         true,
-        "tabs"
+        "tabs",
+        4
       ],
       "no-duplicate-variable": true,
       "no-eval": true,
@@ -64,7 +65,6 @@
         "check-separator",
         "check-type"
       ],
-      "jsdoc-format": true,
       "no-consecutive-blank-lines": true,
       "one-variable-per-declaration": [
         true,
